* tests/factories.py: Moved to goods
