+++ tests/test_api.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from tests import factories
+from goods import factories
 
 User = get_user_model()
